backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai>=1.107.1,<2.0.0
mcp[cli]==1.15.0
openai-agents==0.3.3
python-dotenv==1.1.1
pydantic==2.12.3
